removed pywin32 and pywinty from txt
The pywin32 25 module provides access to many of the Windows APIs from Python. Apps deployed to Streamlit Sharing are executed in a Linux environment 6. As such, pywin32 fails to install on non-Windows systems, including Streamlit Sharing.

To fix the error, remove both pywin32 and pywinpty from lines 72 and 73 57 of your requirements.txt file.